requirements.txt
# SPDX-FileCopyrightText: 2024 Sascha Brawer <sascha@brawer.ch>
# SPDX-License-Identifier: MIT

fudgeo
matplotlib
numpy
opencv-python
pdf2image
Pillow
rasterio
shapely
